add nextron configuration for main thread env args
