+++ renderer/next.config.js
@@ -1,19 +1,7 @@
-const path = require('path');
+const webpack = require('webpack'); // shipped with NextJS
 const withCSS = require('@zeit/next-css');
 const withImages = require('next-images');
-const webpack = require('webpack'); // shipped with NextJS
-
-require('dotenv').config({
-  path: path.join(__dirname, '..', '.env'),
-});
-
-const requiredArgs = ['SENDGRID_TOKEN', 'GITHUB_TOKEN', 'GITHUB_USER'];
-if (requiredArgs.some(arg => !process.env[arg])) {
-  console.error(`Not all required environment variables have been specified.
-Please inspect \`.env.example\`. Required are:
-${JSON.stringify(requiredArgs, null, 2)}`);
-  process.exit(1);
-}
+const { getArgs } = require('../args');
 
 module.exports = withCSS(
   withImages({
@@ -24,15 +12,7 @@ module.exports = withCSS(
 
     webpack: config => {
       config.target = 'electron-renderer';
-
-      const env = requiredArgs.reduce(
-        (acc, curr) => ({
-          ...acc,
-          [`process.env.${curr}`]: JSON.stringify(process.env[curr]),
-        }),
-        {}
-      );
-      config.plugins.push(new webpack.DefinePlugin(env));
+      config.plugins.push(new webpack.DefinePlugin(getArgs()));
 
       return config;
     },
